requirement.txt
langchain_openai
openai
streamlit
huggingface_hub
python-dotenv


